Improve the signup process by checking the username on profiles.wordpress.org
In case the plugin is used elsewhere than on a WordCamp site, using the WordPress.org username and fetching the user description from there can save some time when creating the speakers post type.
Add a description to signup fields to try to make sure the user is using a WordPress.org username and an email linked with Gravatar.