Try to motivate users to add their biography
When a user description is missing, as soon as the corresponding user will post a new Talk, a feedback message will ask him to fill the description field of the user profile.
As the plugin is using the profiles.wordpress.org description during registration, if this description is defined, it will ask to review it for the first talk a user submitted.
Add unit tests to the function to count talks as it is now possible to use it to get a single user s count