First Integration with WordCamp Post types completed.
As it is always a pain to get speakers bio, the front end profile is used to edit the speaker post type. The display name is the Post title and the description is the Post content. It makes the need for a specific button to create the speaker post type (issue #41) useless.
We simply need to check this speaker post type exists before generating the session post type from the Talk Proposal admin UI.